fix bloom file path
