+++ colossalai/inference/tensor_parallel/modeling/__init__.py
@@ -1,0 +1,3 @@
+from .bloom import BloomInferenceForwards
+
+__all__ = ['BloomInferenceForwards']
