+++ colossalai/inference/tensor_parallel/policies/__init__.py
@@ -1,0 +1,3 @@
+from .bloom import BloomModelInferPolicy
+
+__all__ = ['BloomModelInferPolicy']
